requirements.txt
requests==2.26.0
Django==3.2
djangorestframework-simplejwt==5.2.2
django-filter==22.1
asgiref
djangorestframework==3.12.4
gunicorn==20.0.4
psycopg2-binary==2.8.6
PyJWT==2.1.0
pytz==2020.1
sqlparse==0.3.1
webcolors==1.11.1
Pillow==9.0.0
djoser==2.1.0
# django-filter==2.4.0
# Django==2.2.16
# djangorestframework-simplejwt==4.8.0